Moved repeated code to new hook useCategorizeRecentActivity
