+++ src/custom/components/OrdersPanel/index.tsx
@@ -1,12 +1,10 @@
-import { useMemo } from 'react'
 import styled from 'styled-components/macro'
 import { ReactComponent as Close } from 'assets/images/x.svg'
 import AccountDetails from 'components/AccountDetails'
-import useRecentActivity, { TransactionAndOrder } from 'hooks/useRecentActivity'
 import { useWalletInfo } from 'hooks/useWalletInfo'
-import { OrderStatus } from 'state/orders/actions'
 import { useToggleWalletModal } from 'state/application/hooks'
 import { transparentize } from 'polished'
+import { useCategorizeRecentActivity } from '@cow/common/hooks/useCategorizeRecentActivity'
 
 const SideBar = styled.div`
   display: flex;
@@ -126,25 +124,6 @@ const Wrapper = styled.div`
   width: 100%;
 `
 
-const PENDING_STATES = [
-  OrderStatus.PENDING,
-  OrderStatus.PRESIGNATURE_PENDING,
-  OrderStatus.CREATING,
-  OrderStatus.REFUNDING,
-]
-
-const isPending = (data: TransactionAndOrder) => PENDING_STATES.includes(data.status)
-
-const CONFIRMED_STATES = [
-  OrderStatus.FULFILLED,
-  OrderStatus.EXPIRED,
-  OrderStatus.CANCELLED,
-  OrderStatus.REJECTED,
-  OrderStatus.REFUNDED,
-]
-
-const isConfirmed = (data: TransactionAndOrder) => CONFIRMED_STATES.includes(data.status)
-
 export interface OrdersPanelProps {
   handleCloseOrdersPanel: () => void
 }
@@ -153,28 +132,7 @@ export default function OrdersPanel({ handleCloseOrdersPanel }: OrdersPanelProps
   const walletInfo = useWalletInfo()
   const toggleWalletModal = useToggleWalletModal()
 
-  // Returns all RECENT (last day) transaction and orders in 2 arrays: pending and confirmed
-  const allRecentActivity = useRecentActivity()
-
-  const [pendingActivity, confirmedActivity] = useMemo(
-    () =>
-      // Separate the array into 2: transitory (pending) and final (confirmed) states
-      allRecentActivity.reduce<[string[], string[]]>(
-        (acc, activity) => {
-          if (isPending(activity)) {
-            acc[0].push(activity.id)
-          } else if (isConfirmed(activity)) {
-            acc[1].push(activity.id)
-          }
-          return acc
-        },
-        [[], []]
-      ),
-
-    // Reducing unnecessary re-renders
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [JSON.stringify(allRecentActivity)]
-  )
+  const { pendingActivity, confirmedActivity } = useCategorizeRecentActivity()
 
   const { active, ensName } = walletInfo
   const ENSName = ensName
